Add lift and retract options to material tab.
Add exposure, timing, lift and retract specific options to SLAMaterialConfig and the material tab.
These parameters are needed for generating .goo output files and were
not yet present in the slicer. The lift and retract options are
prefixed with 'sla_' to avoid conflict with e.g. 'retract_speed' from
FFF printing.